Quitada implementacion privada
