+++ vsprojects/PTSD-Core/PTSD.vcxproj
@@ -15,7 +15,6 @@
     <ClCompile Include="..\..\source\Core\src\main.cpp" />
     <ClCompile Include="..\..\source\Core\src\PTSDVectors.cpp" />
     <ClCompile Include="..\..\source\Graphics\source\Camera.cpp" />
-    <ClCompile Include="..\..\source\Graphics\source\GraphicsImplementation.cpp" />
     <ClCompile Include="..\..\source\Graphics\source\GraphicsManager.cpp" />
     <ClCompile Include="..\..\source\Input\source\InputImplementation.cpp" />
     <ClCompile Include="..\..\source\Input\source\InputManager.cpp" />
@@ -39,7 +38,6 @@
     <ClInclude Include="..\..\source\Core\include\PTSDVectors.h" />
     <ClInclude Include="..\..\source\Core\include\test.h" />
     <ClInclude Include="..\..\source\Graphics\include\Camera.h" />
-    <ClInclude Include="..\..\source\Graphics\include\GraphicsImplementation.h" />
     <ClInclude Include="..\..\source\Graphics\include\GraphicsLogger.hpp" />
     <ClInclude Include="..\..\source\Graphics\include\GraphicsManager.h" />
     <ClInclude Include="..\..\source\Input\include\InputManager.h" />
